Compression for all file types done: - In createEncoding function I was passing target(char) and it was being compared with node.pair.data(uint8) : Fixed now.

+++ src/DataStructures/data_array.h
@@ -3,12 +3,12 @@
 #include <stdlib.h>
 
 struct Data_array{
-    int *data;
+    uint8_t *data;
     size_t count;
     size_t capacity;
 };
 
-void addDataInDataArray(struct Data_array *data_array, int data){
+void addDataInDataArray(struct Data_array *data_array, uint8_t data){
     if(data_array->count == data_array->capacity){
         if(data_array->capacity == 0){
             data_array->capacity = 1;
@@ -17,7 +17,7 @@ void addDataInDataArray(struct Data_array *data_array, int data){
             data_array->capacity = data_array->capacity*2;
         }
 
-        int *newData = realloc(data_array->data, data_array->capacity * (sizeof(int)));
+        uint8_t *newData = realloc(data_array->data, data_array->capacity * sizeof(uint8_t));
         if(newData == NULL){
             return;
         }
